chakra_apps/stable_diffusion/mac_requirements.txt
reflex==0.3.8
torch
diffusers
pillow
transformers